# docs/swagger.yaml
basePath: /swagger/eco-service
definitions:
...
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
  entity.LabelData:
    properties:
      hb:
        type: number
      id:
        type: string
      label:
        type: string
      tb:
        type: number
      value:
        type: number
    type: object
...
info:
  contact: {}
  description: eco-service API
  title: eco-service API
  version: "1.0"
paths:
  /dashboard/building-floor-power-consumption:
    get:
      description: 建筑楼层用电量，根据粒度获取建筑所有楼层的用电量
      parameters:
      - description: period, hour/day/month/year
        in: query
        name: period
        required: true
        type: string
      - description: building_id
        in: query
        name: building_id
        required: true
        type: string
      - description: query_time,格式2024-01-01,不传则默认当天
        in: query
        name: query_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.LabelData'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 建筑楼层用电量
      tags:
      - Dashboard
  /dashboard/building-power-consumption:
    get:
      description: 建筑用电量，根据粒度获取所有建筑的用电量
      parameters:
      - description: period, hour/day/month/year
        in: query
        name: period
        required: true
        type: string
      - description: query_time,格式2024-01-01,不传则默认当天
        in: query
        name: query_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.LabelData'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 建筑用电量
      tags:
      - Dashboard
  /dashboard/building-type-power-consumption:
    get:
      description: 建筑细分用电量，根据粒度,和分类获取所有建筑的用电量。分类为1:照明，2:动力
      parameters:
      - description: period, hour/day/month/year
        in: query
        name: period
        required: true
        type: string
      - description: query_time,格式2024-01-01,不传则默认当天
        in: query
        name: query_time
        type: string
      - description: type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.LabelData'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 建筑细分用电量
      tags:
      - Dashboard
  /dashboard/park-carbon-emission:
    get:
      description: 园区碳排放，根据粒度获取园区历史的碳排放,不同粒度返回不同数量的数据。日粒度，返回24小时数据。月粒度，返回31天数据。年粒度，返回12个月数据。包括同比环比
      parameters:
      - description: period, hour/day/month/year
        in: query
        name: period
        required: true
        type: string
      - description: query_time,格式2024-01-01,不传则默认当天
        in: query
        name: query_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.LabelData'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 园区碳排放
      tags:
      - Dashboard
  /dashboard/park-power-consumption:
    get:
      description: 园区用电量，根据粒度获取园区历史的用电量,不同粒度返回不同数量的数据。日粒度，返回24小时数据。月粒度，返回31天数据。年粒度，返回12个月数据。包括同比环比
      parameters:
      - description: period, hour/day/month/year
        in: query
        name: period
        required: true
        type: string
      - description: type,1:照明,2:动力
        in: query
        name: type
        type: string
      - description: query_time,格式2024-01-01,不传则默认当天
        in: query
        name: query_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 园区用电量
      tags:
      - Dashboard
  /dashboard/park-standard-coal-emission:
    get:
      description: 园区标准煤排放，根据粒度获取园区历史的标准煤排放,不同粒度返回不同数量的数据。日粒度，返回24小时数据。月粒度，返回31天数据。年粒度，返回12个月数据。包括同比环比
      parameters:
      - description: period, hour/day/month/year
        in: query
        name: period
        required: true
        type: string
      - description: query_time,格式2024-01-01,不传则默认当天
        in: query
        name: query_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.LabelData'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 园区标准煤排放
      tags:
      - Dashboard
  /dashboard/park-water-consumption:
    get:
      description: 园区用水量，根据粒度获取园区当前的用水量,不同粒度返回不同的数据。
      parameters:
      - description: period, hour/day/month/year
        in: query
        name: period
        required: true
        type: string
      - description: query_time,格式2024-01-01,不传则默认当天
        in: query
        name: query_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 园区用水量
      tags:
      - Dashboard
  /dashboard/park-water-consumption-range:
    get:
      description: 园区用水量范围，根据粒度获取园区历史的用水量范围,不同粒度返回不同数量的数据。日粒度，返回24小时数据。月粒度，返回31天数据。年粒度，返回12个月数据。包括同比环比
      parameters:
      - description: period, hour/day/month/year
        in: query
        name: period
        required: true
        type: string
      - description: query_time,格式2024-01-01,不传则默认当天
        in: query
        name: query_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: 园区用水量范围
      tags:
      - Dashboard
...
  /manu_collect:
    get:
      description: Manually collect data
      parameters:
      - description: Start time (2024-01-01)
        in: query
        name: start
        type: string
      - description: End time (2024-01-01)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/common.Response'
      summary: Manually collect data
      tags:
      - Manually collect data
  /manu_gen_demo_water_data:
    get:
      description: Manually generate demo water data
      parameters:
      - description: Start time (2024-01-01)
        in: query
        name: start
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/common.Response'
      summary: Manually generate demo water data
      tags:
      - Manually generate demo water data
swagger: "2.0"